pyproject.toml
[project]
name = "youtubedl"
version = "0.1.0"
description = ""
authors = [
    {name = "TheBiemGamer",email = "47662596+TheBiemGamer@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "yt-dlp (>=2025.1.26,<2026.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "gevent (>=24.11.1,<25.0.0)",
    "dnspython (>=2.7.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

